requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==2.2.3
Flask-JWT-Extended==4.5.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQLdb==2.0.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mysqlclient==2.2.6
PyJWT==2.10.0
PyMySQL==1.1.1
python-dotenv==0.19.2
setuptools==75.5.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==2.2.3
WTForms==3.2.1
